+++ apiphany-core/src/test/java/org/apiphany/utils/security/DeterministicSecureRandom.java
@@ -32,5 +32,4 @@ public class DeterministicSecureRandom extends SecureRandom {
 	public void nextBytes(final byte[] bytes, final SecureRandomParameters params) {
 		nextBytes(bytes);
 	}
-
 }