+++ apiphany-integration-tests/src/test/java/org/apiphany/security/oath2/ApiClientWithOAuth2IT.java
@@ -10,7 +10,6 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import java.time.Duration;
 import java.util.Map;
 
 import org.apiphany.ApiClient;
@@ -22,11 +21,9 @@ import org.apiphany.http.HttpException;
 import org.apiphany.json.JsonBuilder;
 import org.apiphany.lang.ScopedResource;
 import org.apiphany.lang.Strings;
-import org.apiphany.net.Sockets;
 import org.apiphany.security.AuthenticationToken;
 import org.apiphany.security.AuthenticationTokenProvider;
 import org.apiphany.security.AuthenticationType;
-import org.apiphany.security.JwtTokenValidator;
 import org.apiphany.security.oauth2.ClientAuthenticationMethod;
 import org.apiphany.security.oauth2.OAuth2ClientRegistration;
 import org.apiphany.security.oauth2.OAuth2Properties;
@@ -34,9 +31,7 @@ import org.apiphany.security.oauth2.OAuth2ProviderDetails;
 import org.apiphany.security.oauth2.client.OAuth2ApiClient;
 import org.apiphany.security.oauth2.client.OAuth2HttpExchangeClient;
 import org.apiphany.security.oauth2.client.OAuth2HttpExchangeClientBuilder;
-import org.apiphany.security.oauth2.server.SimpleHttpServer;
-import org.apiphany.security.oauth2.server.SimpleOAuth2Server;
-import org.junit.jupiter.api.AfterAll;
+import org.apiphany.security.oauth2.server.JavaSunHttpServer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.morphix.lang.JavaObjects;
@@ -46,21 +41,10 @@ import org.morphix.lang.JavaObjects;
  *
  * @author Radu Sebastian LAZIN
  */
-class ApiClientWithOAuth2IT {
+class ApiClientWithOAuth2IT extends ITWithJavaSunOAuth2Server {
 
-	private static final String CLIENT_SECRET = "apiphany-client-secret-more-than-32-characters";
-	private static final String CLIENT_ID = "apiphany-client";
 	private static final String PROVIDER_NAME = "my-provider-name";
 	private static final String MY_SIMPLE_APP = "my-simple-app";
-
-	private static final Duration PORT_CHECK_TIMEOUT = Duration.ofMillis(500);
-	private static final int OAUTH_SERVER_PORT = Sockets.findAvailableTcpPort(PORT_CHECK_TIMEOUT);
-	private static final int API_SERVER_PORT = Sockets.findAvailableTcpPort(PORT_CHECK_TIMEOUT);
-
-	private static final SimpleOAuth2Server OAUTH2_SERVER = new SimpleOAuth2Server(OAUTH_SERVER_PORT, CLIENT_ID, CLIENT_SECRET);
-	private static final JwtTokenValidator JWT_TOKEN_VALIDATOR = new JwtTokenValidator(CLIENT_ID, CLIENT_SECRET, OAUTH2_SERVER.getUrl());
-
-	private static final SimpleHttpServer API_SERVER = new SimpleHttpServer(API_SERVER_PORT, JWT_TOKEN_VALIDATOR);
 
 	private OAuth2ClientRegistration clientRegistration;
 	private OAuth2ProviderDetails providerDetails;
@@ -70,6 +54,7 @@ class ApiClientWithOAuth2IT {
 	private OAuth2Properties oAuth2Properties;
 
 	@BeforeEach
+	@SuppressWarnings("resource")
 	void setUp() {
 		String clientRegistrationJson = Strings.fromFile("security/oauth2/oauth2-client-registration.json");
 		clientRegistration = JsonBuilder.fromJson(clientRegistrationJson, OAuth2ClientRegistration.class);
@@ -77,18 +62,12 @@ class ApiClientWithOAuth2IT {
 		clientRegistration.setClientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_POST);
 
 		providerDetails = JsonBuilder.fromJson(Strings.fromFile("security/oauth2/oauth2-provider-details.json"), OAuth2ProviderDetails.class);
-		providerDetails.setTokenUri(OAUTH2_SERVER.getUrl() + "/token");
+		providerDetails.setTokenUri(oAuth2Server().getUrl() + "/token");
 
 		oAuth2Properties = OAuth2Properties.of(Map.of(MY_SIMPLE_APP, clientRegistration), Map.of(PROVIDER_NAME, providerDetails));
 
 		clientProperties = new ClientProperties();
 		clientProperties.setCustomProperties(oAuth2Properties);
-	}
-
-	@AfterAll
-	static void cleanup() throws Exception {
-		OAUTH2_SERVER.close();
-		API_SERVER.close();
 	}
 
 	@Test
@@ -99,7 +78,7 @@ class ApiClientWithOAuth2IT {
 			result = client.getName();
 		}
 
-		assertThat(result, equalTo(SimpleHttpServer.NAME));
+		assertThat(result, equalTo(JavaSunHttpServer.NAME));
 
 		verify(client.exchangeClient).close();
 		verify(client.tokenClient).close();
@@ -113,7 +92,7 @@ class ApiClientWithOAuth2IT {
 			result = managedApiClientWithOAuth2.getName();
 		}
 
-		assertThat(result, equalTo(SimpleHttpServer.NAME));
+		assertThat(result, equalTo(JavaSunHttpServer.NAME));
 	}
 
 	@Test
@@ -163,7 +142,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v1ApiClient client = new OAuth2v1ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -172,7 +151,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v2ApiClient client = new OAuth2v2ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -181,7 +160,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v3ApiClient client = new OAuth2v3ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -190,7 +169,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v4ApiClient client = new OAuth2v4ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -210,7 +189,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v5ApiClient client = new OAuth2v5ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -232,7 +211,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v7ApiClient client = new OAuth2v7ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -270,7 +249,7 @@ class ApiClientWithOAuth2IT {
 		try (OAuth2v10ApiClient client = new OAuth2v10ApiClient(clientProperties)) {
 			String result = client.getName();
 
-			assertThat(result, equalTo(SimpleHttpServer.NAME));
+			assertThat(result, equalTo(JavaSunHttpServer.NAME));
 		}
 	}
 
@@ -313,7 +292,7 @@ class ApiClientWithOAuth2IT {
 			try (OAuth2UnmanagedTokenClientApiClient client = new OAuth2UnmanagedTokenClientApiClient(clientProperties, tokenClient)) {
 				String result = client.getName();
 
-				assertThat(result, equalTo(SimpleHttpServer.NAME));
+				assertThat(result, equalTo(JavaSunHttpServer.NAME));
 			}
 
 			verify(tokenClient, times(0)).close();
@@ -355,11 +334,12 @@ class ApiClientWithOAuth2IT {
 			super(exchangeClientResource);
 		}
 
+		@SuppressWarnings("resource")
 		public String getName() {
 			return client()
 					.http()
 					.get()
-					.url("http://localhost:" + API_SERVER_PORT)
+					.url("http://localhost:" + apiServer().getPort())
 					.path(API, "name")
 					.retrieve(String.class)
 					.orNull();