+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `X25519Keys` key exchange handler to handle `XDH` keys.
 - Moved `OAuth2ApiClientIT` integration test and `testcontainers` dependencies to `apiphany-integration-test` including relevant resource files.
 - Added `Header.value` convenience static methods to build a header value which simply delegates to their equivalent `HeaderValues.value`.
+- Moved `AbstractHttpExchangeClient.getTracingHeaders` method to `HttpExchangeClient` interface as a default method.
 
 ---